+++ pom.xml
@@ -2,8 +2,8 @@
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.comp</groupId>
-  <artifactId>proj</artifactId>
+  <groupId>com.lightning.lambdas</groupId>
+  <artifactId>UserExporter</artifactId>
   <version>1.0-SNAPSHOT</version>
   <dependencies>
     <dependency>
