--- src/main/java/com/comp/Unit_tioih_root_1_1.java
@@ -1,11 +1,0 @@
-package com.comp;
-
-import com.comp.aws.dynamo.DynamoTable;
-import com.comp.config.Config;
-
-public class Unit_tioih_root_1_1 {
-  public DynamoTable execute(final Config config) {
-
-    return config.usersTable;
-  }
-}
