--- src/main/java/com/comp/Unit_tioih_root_1_3.java
@@ -1,11 +1,0 @@
-package com.comp;
-
-import com.amazonaws.services.dynamodbv2.document.Item;
-
-public class Unit_tioih_root_1_3 {
-  public String execute(final Item item) {
-
-    String userJson = item.toJSON();
-    return userJson;
-  }
-}
