--- src/main/java/com/comp/Unit_tioih_root_2_1.java
@@ -1,11 +1,0 @@
-package com.comp;
-
-import com.comp.aws.s3.S3Bucket;
-import com.comp.config.Config;
-
-public class Unit_tioih_root_2_1 {
-  public S3Bucket execute(final Config config) {
-
-    return config.usersBucket;
-  }
-}
