--- src/main/java/com/comp/UserIdParser.java
@@ -1,9 +1,0 @@
-package com.comp;
-
-public class UserIdParser {
-  public String execute(final String fileName) {
-
-    String userId = fileName.split(":")[0];
-    return userId;
-  }
-}
