--- src/main/java/com/comp/aws/s3/S3FileWriter.java
@@ -1,13 +1,0 @@
-package com.comp.aws.s3;
-
-public class S3FileWriter {
-  private final S3ClientProvider s3ClientProvider;
-
-  public S3FileWriter(final S3ClientProvider s3ClientProvider) {
-    this.s3ClientProvider = s3ClientProvider;
-  }
-
-  public void execute(final S3Bucket s3Bucket, final String path, final String content) {
-    s3ClientProvider.get().putObject(s3Bucket.bucketName, path, content);
-  }
-}
