--- src/main/java/com/comp/config/EnvironmentProvider.java
@@ -1,7 +1,0 @@
-package com.comp.config;
-
-public class EnvironmentProvider {
-  public String execute() {
-    return System.getenv("Environment");
-  }
-}
