+++ src/main/java/com/lightning/lambdas/DynamoTableProvider.java
@@ -1,0 +1,11 @@
+package com.lightning.lambdas;
+
+import com.lightning.lambdas.aws.dynamo.DynamoTable;
+import com.lightning.lambdas.config.Config;
+
+public class DynamoTableProvider {
+  public DynamoTable execute(final Config config) {
+
+    return config.usersTable;
+  }
+}
