+++ src/main/java/com/lightning/lambdas/S3BucketProvider.java
@@ -1,0 +1,11 @@
+package com.lightning.lambdas;
+
+import com.lightning.lambdas.aws.s3.S3Bucket;
+import com.lightning.lambdas.config.Config;
+
+public class S3BucketProvider {
+  public S3Bucket execute(final Config config) {
+
+    return config.usersBucket;
+  }
+}
