+++ src/main/java/com/lightning/lambdas/S3UserWriter.java
@@ -1,0 +1,29 @@
+package com.lightning.lambdas;
+
+import com.lightning.lambdas.aws.s3.S3Bucket;
+import com.lightning.lambdas.aws.s3.S3FileWriter;
+import com.lightning.lambdas.config.Config;
+import com.lightning.lambdas.config.ConfigProvider;
+
+public class S3UserWriter {
+  private final ConfigProvider configProvider;
+
+  private final S3BucketProvider s3BucketProvider;
+
+  private final S3FileWriter s3Writer;
+
+  public S3UserWriter(
+      final ConfigProvider configProvider,
+      final S3BucketProvider s3BucketProvider,
+      final S3FileWriter s3Writer) {
+    this.configProvider = configProvider;
+    this.s3BucketProvider = s3BucketProvider;
+    this.s3Writer = s3Writer;
+  }
+
+  public void execute(final String path, final String content) {
+    final Config config = configProvider.execute();
+    final S3Bucket s3Bucket = s3BucketProvider.execute(config);
+    s3Writer.execute(s3Bucket, path, content);
+  }
+}
