+++ src/main/java/com/lightning/lambdas/UserIdParser.java
@@ -1,0 +1,9 @@
+package com.lightning.lambdas;
+
+public class UserIdParser {
+  public String execute(final String fileName) {
+
+    String userId = fileName.split("-")[0];
+    return userId;
+  }
+}
