+++ src/main/java/com/lightning/lambdas/config/EnvironmentProvider.java
@@ -1,0 +1,7 @@
+package com.lightning.lambdas.config;
+
+public class EnvironmentProvider {
+  public String execute() {
+    return System.getenv("Environment");
+  }
+}
