+++ src/main/resources/application.conf
@@ -1,0 +1,15 @@
+
+scratch {
+
+    usersBucket {
+        bucketName = ${?scratch_usersBucket_bucketName}
+    }
+
+
+    usersTable {
+        region = ${?scratch_usersTable_region}
+        tableName = ${?scratch_usersTable_tableName}
+        idFieldName = ${?scratch_usersTable_idFieldName}
+    }
+
+}
