--- src/test/java/com/comp/aws/s3/S3FileWriterTest.java
@@ -1,30 +1,0 @@
-package com.comp.aws.s3;
-
-import org.junit.jupiter.api.BeforeEach;
-
-class S3FileWriterTest {
-
-  // Values
-  private static S3Bucket s3Bucket;
-  private static String content;
-  private static String fileName;
-
-  static {
-    s3Bucket = new S3Bucket("");
-    content = "some text content";
-    fileName = "/a/file/path";
-  }
-
-  // Test Fixture
-  private S3FileWriter s3FileWriter;
-
-  @BeforeEach
-  void setupTestFixture() {
-    s3FileWriter = new S3FileWriter(new S3ClientProvider());
-  }
-
-  // @Test - runs against AWS S3
-  void test1() {
-    s3FileWriter.execute(s3Bucket, fileName, content);
-  }
-}
